Several refactorings
- Load sounds based on sound metadata not files on disk
- Don't download if file already exists
- Support more query parameters and midi offset control

+++ SourceSampler/Source/PluginProcessor.cpp
@@ -27,15 +27,16 @@ SourceSamplerAudioProcessor::SourceSamplerAudioProcessor()
 #endif
 {
     tmpDownloadLocation = File::getSpecialLocation(File::userDocumentsDirectory).getChildFile("FreesoundSimpleSampler");
-    tmpDownloadLocation.deleteRecursively();
-    tmpDownloadLocation.createDirectory();
+    
     midicounter = 1;
     startTime = Time::getMillisecondCounterHiRes() * 0.001;
     
     // Start with a default random query
     std::vector<String> queries = {"wood", "metal", "glass", "percussion", "cat", "hit", "drums"};
     auto randomInt = juce::Random::getSystemRandom().nextInt(queries.size());
-    makeQueryAndLoadSounds(queries[randomInt]);
+    int numSounds = 16;
+    float maxSoundLength = 0.5;
+    makeQueryAndLoadSounds(queries[randomInt], numSounds, maxSoundLength);
     
     // Configure processor to listen messages from server interface
     serverInterface.addActionListener(this);
@@ -44,9 +45,13 @@ SourceSamplerAudioProcessor::SourceSamplerAudioProcessor()
 SourceSamplerAudioProcessor::~SourceSamplerAudioProcessor()
 {
     // Deletes the tmp directory so downloaded files do not stay there
-    tmpDownloadLocation.deleteRecursively();
-    for (int i = 0; i < downloadTasksToDelete.size(); i++) {
-        downloadTasksToDelete.at(i).reset();
+    if (tmpDownloadLocation.exists()){
+        tmpDownloadLocation.deleteRecursively();
+    }
+    
+    // Delete download task objects
+    for (int i = 0; i < downloadTasks.size(); i++) {
+        downloadTasks.at(i).reset();
     }
     
     // Remove this as a listener from serverInterface
@@ -190,17 +195,29 @@ void SourceSamplerAudioProcessor::setStateInformation (const void* data, int siz
 
 void SourceSamplerAudioProcessor::actionListenerCallback (const String &message)
 {
+    std::cout << "Action message: " << message << std::endl;
+    
     if (message.startsWith(String(ACTION_NEW_QUERY_TRIGGERED_FROM_SERVER))){
+        String serializedParameters = message.substring(String(ACTION_NEW_QUERY_TRIGGERED_FROM_SERVER).length() + 1);
+        StringArray tokens;
+        tokens.addTokens (serializedParameters, (String)SERIALIZATION_SEPARATOR, "");
+        String query = tokens[0];
+        int numSounds = tokens[1].getIntValue();
+        float maxSoundLength = tokens[2].getFloatValue();
+        std::cout << "New query triggered from server: " << query << " " << numSounds << " " << maxSoundLength << std::endl;
+        makeQueryAndLoadSounds(query, numSounds, maxSoundLength);
         
-        String query = message.substring(String(ACTION_NEW_QUERY_TRIGGERED_FROM_SERVER).length() + 1);
-        std::cout << "New query triggered from server: " << query << std::endl;
-        makeQueryAndLoadSounds(query);
+    } else if (message.startsWith(String(ACTION_SET_MIDI_ROOT_NOTE_OFFSET))){
+        int newOffset = message.substring(String(ACTION_SET_MIDI_ROOT_NOTE_OFFSET).length() + 1).getIntValue();
+        std::cout << "Offsetting midi root note by: " << newOffset << std::endl;
+        setSources(newOffset);
+        
     }
 }
 
 //==============================================================================
 
-void SourceSamplerAudioProcessor::makeQueryAndLoadSounds(const String& query)
+void SourceSamplerAudioProcessor::makeQueryAndLoadSounds(const String& query, int numSounds, float maxSoundLength)
 {
     if (isQueryinAndDownloadingSounds){
         std::cout << "Source is already busy querying and downloading sounds" << std::endl;
@@ -209,20 +226,27 @@ void SourceSamplerAudioProcessor::makeQueryAndLoadSounds(const String& query)
     std::cout << "Querying new sounds for: " << query << std::endl;
     isQueryinAndDownloadingSounds = true;
     FreesoundClient client(FREESOUND_API_KEY);
-    SoundList list = client.textSearch(query, "duration:[0 TO 0.5]", "score", 1, -1, 150, "id,name,username,license,previews");
-    Array<FSSound> sounds = list.toArrayOfSounds();
-    std::random_shuffle(sounds.begin(), sounds.end());
-    sounds.resize(16);
-    std::vector<juce::StringArray> soundsInfo;
-    for (int i=0; i<sounds.size(); i++){
-        FSSound sound = sounds[i];
-        StringArray soundData;
-        soundData.add(sound.name);
-        soundData.add(sound.user);
-        soundData.add(sound.license);
-        soundsInfo.push_back(soundData);
+    auto filter = "duration:[0 TO " + (String)maxSoundLength + "]";
+    SoundList list = client.textSearch(query, filter, "score", 1, -1, 150, "id,name,username,license,previews");
+    if (list.getCount() > 0){
+        std::cout << "Query got " << list.getCount() << " results" << std::endl;
+        Array<FSSound> sounds = list.toArrayOfSounds();
+        std::random_shuffle(sounds.begin(), sounds.end());
+        sounds.resize(jmin(numSounds, list.getCount()));
+        std::vector<juce::StringArray> soundsInfo;
+        for (int i=0; i<sounds.size(); i++){
+            FSSound sound = sounds[i];
+            StringArray soundData;
+            soundData.add(sound.id);
+            soundData.add(sound.name);
+            soundData.add(sound.user);
+            soundData.add(sound.license);
+            soundsInfo.push_back(soundData);
+        }
+        newSoundsReady(sounds, query, soundsInfo);
+    } else {
+        std::cout << "Query got no results..." << std::endl;
     }
-    newSoundsReady(sounds, query, soundsInfo);
 }
 
 void SourceSamplerAudioProcessor::newSoundsReady (Array<FSSound> sounds, String textQuery, std::vector<juce::StringArray> soundsInfo)
@@ -231,64 +255,107 @@ void SourceSamplerAudioProcessor::newSoundsReady (Array<FSSound> sounds, String 
     // been made and new sounda have been selected for loading into the sampler.
     // This methods downloads the sounds, sotres in tmp directory and...
     
+    //
+    if (!tmpDownloadLocation.exists()){
+        tmpDownloadLocation.createDirectory();
+    }
+    
     // Download the sounds
-    std::cout << "Downloading new sounds" << std::endl;
+    std::cout << "Downloading new sounds..." << std::endl;
     FreesoundClient client(FREESOUND_API_KEY);
     for (int i=0; i<sounds.size(); i++){
         File location = tmpDownloadLocation.getChildFile(sounds[i].id).withFileExtension("mp3");
-        //std::cout << location.getFullPathName() << std::endl;
-        std::unique_ptr<URL::DownloadTask> downloadTask = client.downloadMP3SoundPreview(sounds[i], location);
-        downloadTasksToDelete.push_back(std::move(downloadTask));
+        if (!location.exists()){  // Dont' re-download if file already exists
+            std::unique_ptr<URL::DownloadTask> downloadTask = client.downloadMP3SoundPreview(sounds[i], location);
+            downloadTasks.push_back(std::move(downloadTask));
+        }
     }
+    
+    std::cout << "Waiting for all download tasks to finish..." << std::endl;
+    bool allFinished = false;
+    while (!allFinished){
+        allFinished = true;
+        for (int i=0; i<downloadTasks.size(); i++){
+            if (!downloadTasks[i]->isFinished()){
+                allFinished = false;
+            }
+        }
+    }
+    
+    // Store info about the query and tell UI component(s) to update
     query = textQuery;
     soundsArray = soundsInfo;
     sendActionMessage(String(ACTION_SHOULD_UPDATE_SOUNDS_TABLE));
     
     // Load the sounds in the sampler
-    std::cout << "Loading downloaded sounds to sampler" << std::endl;
-    setSources();
+    setSources(0);
     
     std::cout << "All done and ready!" << std::endl << std::endl;
     isQueryinAndDownloadingSounds = false;
 }
 
-void SourceSamplerAudioProcessor::setSources()
+void SourceSamplerAudioProcessor::setSources(int midiNoteRootOffset)
 {
+    sampler.clearSounds();
+    sampler.clearVoices();
+    
+    // Configure sampler basics
     int poliphony = 16;
-    int maxLength = 10;
     for (int i = 0; i < poliphony; i++) {
         sampler.addVoice(new SamplerVoice());
     }
-
+    
     if(audioFormatManager.getNumKnownFormats() == 0){
         audioFormatManager.registerBasicFormats();
     }
 
-    Array<File> files = tmpDownloadLocation.findChildFiles(2, false);
-    for (int i = 0; i < files.size(); i++) {
-        std::unique_ptr<AudioFormatReader> reader(audioFormatManager.createReaderFor(files[i]));
-        BigInteger notes;
-        notes.setRange(i * 8, i * 8 + 7, true);
-        sampler.addSound(new SamplerSound(String(i), *reader, notes, i*8, 0, maxLength, maxLength));
-        //reader.release();
+    int attackTime = 0;
+    int releaseTime = 1;
+    int maxSampleLength = 20;  // This is unrelated to the maxSoundLength of the makeQueryAndLoadSounds method
+    int nSounds = (int)soundsArray.size();
+    
+    std::cout << "Loading " << nSounds << " sounds to sampler" << std::endl;
+    if (nSounds > 0){
+        int nNotesPerSound = 128 / nSounds;
+        for (int i = 0; i < nSounds; i++) {
+            String soundID = soundsArray[i][0];
+            File audioSample = tmpDownloadLocation.getChildFile(soundID).withFileExtension("mp3");
+            if (audioSample.exists() && audioSample.getSize() > 0){  // Check that file exists and is not empty
+                std::unique_ptr<AudioFormatReader> reader(audioFormatManager.createReaderFor(audioSample));
+                int midiNoteForNormalPitch = i * nNotesPerSound + nNotesPerSound / 2 + midiNoteRootOffset;
+                BigInteger midiNotes;
+                midiNotes.setRange(i * nNotesPerSound, nNotesPerSound, true);
+                std::cout << "- Adding sound " << audioSample.getFullPathName() << " with midi root note " << midiNoteForNormalPitch << std::endl;
+                sampler.addSound(new SamplerSound(String(i), *reader, midiNotes, midiNoteForNormalPitch, attackTime, releaseTime, maxSampleLength));
+                
+            } else {
+                std::cout << "- Skipping sound " << soundID << " (no file found or file is empty)" << std::endl;
+            }
+        }
     }
+    std::cout << "Sampler sources configured with " << sampler.getNumSounds() << " sounds and " << sampler.getNumVoices() << " voices" << std::endl;
 }
 
-void SourceSamplerAudioProcessor::addToMidiBuffer(int notenumber)
+void SourceSamplerAudioProcessor::addToMidiBuffer(int soundNumber)
 {
+    
+    int nSounds = (int)soundsArray.size();
+    if (nSounds > 0){
+        int nNotesPerSound = 128 / nSounds;
+        int midiNoteForNormalPitch = soundNumber * nNotesPerSound + nNotesPerSound / 2;
+        
+        MidiMessage message = MidiMessage::noteOn(1, midiNoteForNormalPitch, (uint8)127);
+        double timestamp = Time::getMillisecondCounterHiRes() * 0.001 - getStartTime();
+        message.setTimeStamp(timestamp);
 
-    MidiMessage message = MidiMessage::noteOn(10, notenumber, (uint8)100);
-    double timestamp = Time::getMillisecondCounterHiRes() * 0.001 - getStartTime();
-    message.setTimeStamp(timestamp);
+        auto sampleNumber = (int)(timestamp * getSampleRate());
 
-    auto sampleNumber = (int)(timestamp * getSampleRate());
+        midiFromEditor.addEvent(message,sampleNumber);
 
-    midiFromEditor.addEvent(message,sampleNumber);
-
-    //auto messageOff = MidiMessage::noteOff(message.getChannel(), message.getNoteNumber());
-    //messageOff.setTimeStamp(Time::getMillisecondCounterHiRes() * 0.001 - startTime);
-    //midiFromEditor.addEvent(messageOff,sampleNumber+1);
-
+        //auto messageOff = MidiMessage::noteOff(message.getChannel(), message.getNoteNumber());
+        //messageOff.setTimeStamp(Time::getMillisecondCounterHiRes() * 0.001 - startTime);
+        //midiFromEditor.addEvent(messageOff,sampleNumber+1);
+    }
 }
 
 double SourceSamplerAudioProcessor::getStartTime(){
